Address bugbot follow-up feedback
