+++ packages/utilities/routes.ts
@@ -1,0 +1,5 @@
+export const normalizeRoutePath = (value: string): string => {
+  if (!value) return '/';
+  const normalized = value === '/' ? value : value.replace(/\/$/, '');
+  return normalized || '/';
+};
